refactore complete code to ve able to create a installable package
