+++ examples/compare_convergence.py
