+++ examples/example_fracture_vpinns.py
