+++ examples/example_int_weak_plus_posterri.py
