+++ examples/example_weak.py
