+++ examples/example_weak_convergence.py
