+++ examples/example_weak_int.py
