+++ examples/example_weak_plus_posterri.py
