+++ examples/seven_fractures/vertices_data.txt
