+++ src/basis/__init__.py
@@ -1,0 +1,7 @@
+"""Initialization of basis classes"""
+
+from .abstract_basis import AbstractBasis
+from .basis import Basis
+from .interior_edges_basis import InteriorEdgesBasis
+from .fracture_basis import FractureBasis
+from .interior_edges_fracture_basis import InteriorEdgesFractureBasis
