+++ src/element/__init__.py
@@ -1,0 +1,5 @@
+"""Initialization of element classes"""
+
+from .abstract_element import AbstractElement
+from .element_line import ElementLine
+from .element_tri import ElementTri
