+++ src/element/element_tri.py
@@ -1,0 +1,143 @@
+import torch
+from .abstract_element import AbstractElement
+
+
+class ElementTri(AbstractElement):
+    """Class for 2D triangular element representation"""
+
+    @property
+    def barycentric_grad(self):
+        return torch.tensor([[-1.0, -1.0], [1.0, 0.0], [0.0, 1.0]])
+
+    @property
+    def reference_element_area(self):
+        return 0.5
+
+    @property
+    def outward_normal(self):
+        """Return the outward normal vectors of the reference triangle edges."""
+        return torch.tensor([[1.0, 1.0], [-1.0, 0.0], [0.0, -1.0]])
+
+    def compute_barycentric_coordinates(self, x: torch.Tensor):
+        return torch.stack(
+            [1.0 - x[..., [0]] - x[..., [1]], x[..., [0]], x[..., [1]]], dim=-2
+        )
+
+    def compute_shape_functions(
+        self, bar_coords: torch.Tensor, inv_map_jacobian: torch.Tensor
+    ):
+        lambda_1, lambda_2, lambda_3 = torch.split(bar_coords, 1, dim=-2)
+
+        grad_lambda_1, grad_lambda_2, grad_lambda_3 = torch.split(
+            self.barycentric_grad, 1, dim=-2
+        )
+
+        if self.polynomial_order == 1:
+
+            v = bar_coords
+
+            v_grad = self.barycentric_grad @ inv_map_jacobian
+
+        elif self.polynomial_order == 2:
+
+            v = torch.concat(
+                [
+                    lambda_1 * (2 * lambda_1 - 1),
+                    lambda_2 * (2 * lambda_2 - 1),
+                    lambda_3 * (2 * lambda_3 - 1),
+                    4 * lambda_1 * lambda_2,
+                    4 * lambda_2 * lambda_3,
+                    4 * lambda_3 * lambda_1,
+                ],
+                dim=-2,
+            )
+
+            v_grad = (
+                torch.concat(
+                    [
+                        (4 * lambda_1 - 1) * grad_lambda_1,
+                        (4 * lambda_2 - 1) * grad_lambda_2,
+                        (4 * lambda_3 - 1) * grad_lambda_3,
+                        4 * (lambda_2 * grad_lambda_1 + lambda_1 * grad_lambda_2),
+                        4 * (lambda_3 * grad_lambda_2 + lambda_2 * grad_lambda_3),
+                        4 * (lambda_1 * grad_lambda_3 + lambda_3 * grad_lambda_1),
+                    ],
+                    dim=-2,
+                )
+                @ inv_map_jacobian
+            )
+        else:
+
+            raise NotImplementedError("Polynomial order not implemented")
+
+        return v, v_grad
+
+    def _compute_gauss_values(self):
+
+        if self.integration_order == 1:
+
+            gaussian_nodes = torch.tensor([[1 / 3, 1 / 3]])
+
+            gaussian_weights = torch.tensor([[[1.0]]])
+
+        elif self.integration_order == 2:
+
+            gaussian_nodes = torch.tensor(
+                [[1 / 6, 1 / 6], [2 / 3, 1 / 6], [1 / 6, 2 / 3]]
+            )
+
+            gaussian_weights = torch.tensor([[[1 / 3]], [[1 / 3]], [[1 / 3]]])
+
+        elif self.integration_order == 3:
+
+            gaussian_nodes = torch.tensor(
+                [[1 / 3, 1 / 3], [0.6, 0.2], [0.2, 0.6], [0.2, 0.2]]
+            )
+
+            gaussian_weights = torch.tensor(
+                [[[-9 / 16]], [[25 / 48]], [[25 / 48]], [[25 / 48]]]
+            )
+
+        elif self.integration_order == 4:
+
+            gaussian_nodes = torch.tensor(
+                [
+                    [0.816847572980459, 0.091576213509771],
+                    [0.091576213509771, 0.816847572980459],
+                    [0.091576213509771, 0.091576213509771],
+                    [0.108103018168070, 0.445948490915965],
+                    [0.445948490915965, 0.108103018168070],
+                    [0.445948490915965, 0.445948490915965],
+                ]
+            )
+
+            gaussian_weights = torch.tensor(
+                [
+                    [[0.109951743655322]],
+                    [[0.109951743655322]],
+                    [[0.109951743655322]],
+                    [[0.223381589678011]],
+                    [[0.223381589678011]],
+                    [[0.223381589678011]],
+                ]
+            )
+        else:
+
+            raise NotImplementedError("Integration order not implemented")
+
+        return gaussian_nodes, gaussian_weights
+
+    def compute_det_and_inv_map(self, map_jacobian: torch.Tensor):
+
+        ab, cd = torch.split(map_jacobian, 1, dim=-2)
+
+        a, b = torch.split(ab, 1, dim=-1)
+        c, d = torch.split(cd, 1, dim=-1)
+
+        det_map_jacobian = (a * d - b * c).unsqueeze(-3)
+
+        inv_map_jacobian = (1 / det_map_jacobian) * torch.stack(
+            [torch.concat([d, -b], dim=-1), torch.concat([-c, a], dim=-1)], dim=-2
+        )
+
+        return det_map_jacobian, inv_map_jacobian
