+++ src/mesh/__init__.py
@@ -1,0 +1,5 @@
+"""Initialization of mesh classes"""
+
+from .abstract_mesh import AbstractMesh
+from .mesh_tri import MeshTri
+from .fractures_tri import FracturesTri
