+++ src/model/__init__.py
@@ -1,0 +1,2 @@
+from .neural_network import FeedForwardNeuralNetwork
+from .model import Model
