+++ src/pytorch_fem_solver.egg-info/PKG-INFO
@@ -1,0 +1,17 @@
+Metadata-Version: 2.4
+Name: pytorch_fem_solver
+Version: 1.0.0
+Summary: A finite element solver for Pytorch.
+Author-email: Nicolas Zamorano <nicolas.zamo12@gmail.com>
+License-Expression: MIT
+Project-URL: Homepage, https://github.com/Nicolas-Zamorano/pytorch_fem_solver
+Project-URL: Issues, https://github.com/Nicolas-Zamorano/pytorch_fem_solver/issues
+Classifier: Programming Language :: Python :: 3
+Classifier: Operating System :: OS Independent
+Requires-Python: >=3.9
+Description-Content-Type: text/markdown
+License-File: LICENSE
+Dynamic: license-file
+
+# pytorch_fem_solver
+FEM solver based on Pytorch
