+++ src/pytorch_fem_solver.egg-info/SOURCES.txt
@@ -1,0 +1,35 @@
+LICENSE
+README.md
+pyproject.toml
+src/__init__.py
+src/basis/__init__.py
+src/basis/abstract_basis.py
+src/basis/basis.py
+src/basis/fracture_basis.py
+src/basis/interior_edges_basis.py
+src/basis/interior_edges_fracture_basis.py
+src/element/__init__.py
+src/element/abstract_element.py
+src/element/element_line.py
+src/element/element_tri.py
+src/mesh/__init__.py
+src/mesh/abstract_mesh.py
+src/mesh/fractures_tri.py
+src/mesh/mesh_tri.py
+src/model/__init__.py
+src/model/model.py
+src/model/neural_network.py
+src/pytorch_fem_solver.egg-info/PKG-INFO
+src/pytorch_fem_solver.egg-info/SOURCES.txt
+src/pytorch_fem_solver.egg-info/dependency_links.txt
+src/pytorch_fem_solver.egg-info/top_level.txt
+tests/test_1_fracture.py
+tests/test_1_fracture_2d.py
+tests/test_assembly.py
+tests/test_derivate_wrt_inputs.py
+tests/test_fracture_jump.py
+tests/test_fracture_map.py
+tests/test_global_local_idx.py
+tests/test_interpolation_to_lower_space.py
+tests/test_jump.py
+tests/test_triangulation_int_space.py
