+++ src/pytorch_fem_solver.egg-info/top_level.txt
@@ -1,0 +1,5 @@
+__init__
+basis
+element
+mesh
+model
