+++ tests/test_assembly.py
