+++ tests/test_global_local_idx.py
